Initial commit of Aaron C. Meadows AaronUPS Version 0.4. from http://www.phpbuilder.com/snippet/detail.php?type=snippet&id=712
Our in house tools have been using this for some time now. Looks like it hasn't been updated since 2005, but we needed some changes made. In the near future I expect to be completely replacing the tools that use this, so didn't seem worth a complete updated rewrite. But, figured sharing updates is sharing.